+++ Tests/Test_activity.py
@@ -1,5 +1,5 @@
 import unittest
-from activities import *
+from activity import *
 
 class MyTestCase(unittest.TestCase):
     def test_something(self):
